requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
pydantic==1.9.1
qrcode==7.3.1
Pillow==9.2.0
flake8==5.0.2
pymemcache==3.5.2
pytest==7.1.2
